requirements.txt
Argparse
Iris
Watchdog
Logging
Monty
Numpy
#OpenGL
PyPng
PySide
#Qt
Requests
Scipy
Vispy